move extractFrom from EcommerceSpider to BaseSpider
